# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter Ops API",
    "description": "Clean REST API adapter for the Urja Meter Ops portal",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Root  Get"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "List Meters",
        "operationId": "list_meters_api_v1_meters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeterSummary"
                  },
                  "type": "array",
                  "title": "Response List Meters Api V1 Meters Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get Meter Detail",
        "operationId": "get_meter_detail_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/consumption": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get Meter Consumption",
        "operationId": "get_meter_consumption_api_v1_meters__meter_id__consumption_get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/geo": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get Meter Geo",
        "operationId": "get_meter_geo_api_v1_meters__meter_id__geo_get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterGeoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConsumptionRecord": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "kwh": {
            "type": "number",
            "title": "Kwh"
          },
          "kvah": {
            "type": "number",
            "title": "Kvah"
          },
          "volt_r": {
            "type": "number",
            "title": "Volt R"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "kwh",
          "kvah",
          "volt_r"
        ],
        "title": "ConsumptionRecord"
      },
      "ConsumptionResponse": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ConsumptionRecord"
            },
            "type": "array",
            "title": "Records",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "meter_id"
        ],
        "title": "ConsumptionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MeterDetail": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "type": "string",
            "title": "Phase Type"
          },
          "install_status": {
            "type": "string",
            "title": "Install Status"
          },
          "install_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Install Type"
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          },
          "hierarchy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeterHierarchy"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status"
        ],
        "title": "MeterDetail"
      },
      "MeterGeoResponse": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "latitude",
          "longitude"
        ],
        "title": "MeterGeoResponse"
      },
      "MeterHierarchy": {
        "properties": {
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          },
          "circle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circle"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "subdivision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subdivision"
          },
          "substation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substation"
          },
          "feeder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feeder"
          },
          "dt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt"
          }
        },
        "type": "object",
        "title": "MeterHierarchy"
      },
      "MeterSummary": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "type": "string",
            "title": "Phase Type"
          },
          "install_status": {
            "type": "string",
            "title": "Install Status"
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status"
        ],
        "title": "MeterSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}